Add security, tests

+++ src/main/java/com/OrderManagement/service/BouquetService.java
@@ -1,5 +1,7 @@
 package com.OrderManagement.service;
 
+import com.OrderManagement.exception.BouquetNotFoundException;
+import com.OrderManagement.exception.UserNotFoundException;
 import com.OrderManagement.model.Bouquet;
 import com.OrderManagement.repository.BouquetRepository;
 import lombok.AllArgsConstructor;
@@ -17,17 +19,26 @@ public class BouquetService {
         log.info("Bouquet was added successfully");
     }
 
-    public Bouquet getBouquetById(Long id) {
-        return bouquetRepository.findById(id).get();
+    public Bouquet getBouquetById(Long id) throws BouquetNotFoundException {
+        return bouquetRepository.findById(id).orElseThrow(() ->
+                new BouquetNotFoundException("Bouquet by id: " + id + " was not found"));
+
     }
 
     public void updateBouquet(Bouquet bouquet) {
-            bouquetRepository.save(bouquet);
-            log.info("Bouquet was updated successfully");
+        bouquetRepository.save(bouquet);
+        log.info("Bouquet was updated successfully");
     }
 
-    public void deleteBouquetById(Long id) {
+    public void deleteBouquetById(Long id) throws BouquetNotFoundException {
+        bouquetRepository.deleteById(id);
+        log.info("Bouquet was deleted successfully");
+
+        if (bouquetRepository.existsById(id)) {
             bouquetRepository.deleteById(id);
             log.info("Bouquet was deleted successfully");
+        } else {
+            throw new BouquetNotFoundException("Bouquet by id: " + id + " was not found");
+        }
     }
 }
